requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
clarifai-grpc==7.11.0
click==8.1.7
dnspython==2.6.1
emoji==2.12.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
idna==3.7
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
protobuf==3.19.6
pymongo==4.7.3
python-telegram-bot==13.15
pytz==2024.1
requests==2.32.3
six==1.16.0
tornado==6.1
typing_extensions==4.12.0
tzlocal==5.2
urllib3==2.2.1
